+++ rear/user/src/main/java/com/spaceobj/user/bo/SysPhotoBo.java
@@ -17,7 +17,7 @@ import java.io.Serializable;
 public class SysPhotoBo implements Serializable {
 
   /** 图片id */
-  private Long photoId;
+  private Integer photoId;
 
   /** 图片URl */
   private String photoUrl;
